+++ src/pairag/mcp/prompts.py
@@ -90,6 +90,3 @@ PROMPT_WITHOUT_DEEP_RESEARCH = """
   - 对于违反法律、道德和社会责任的非法行为以及涉及不安全/不道德/敏感/色情/暴力/赌博/违法等行为的问题，或者是可能存在非法目的的问题，请明确拒绝提供所要求的信息，并简单解释为什么这样的请求不能被满足。
 6. 请仔细思考后，给出你的答案：
 """
-
-
-DEFAULT_SUMMARIZER_PROMPT = "请根据上下文对下面这段文字生成详细的总结"
